TransactionDraft to interface. #942
